WIP create postUserid.ts
ユーザーのIDを取得する
